fix i18n post-refactor issues
